requirements.txt
flask-cors
mysql-connector-python